Legger til K9Beskjed. Utvider cleanupstream til å produsere til to ulike topics, k9.rapid og k9-dittnav-varsel
