+++ src/main/kotlin/no/nav/helse/prosessering/v1/asynkron/Topics.kt
@@ -2,7 +2,6 @@ package no.nav.helse.prosessering.v1.asynkron
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.PropertyNamingStrategy
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
@@ -53,6 +52,11 @@ internal object Topics {
         name = "omsorgspenger.k9-rapid-v2",
         serDes = SerDes()
     )
+
+    val K9_DITTNAV_VARSEL = Topic(
+        name = "dusseldorf.privat-k9-dittnav-varsel-beskjed",
+        serDes = SerDes()
+    )
 }
 
 internal fun TopicEntry.deserialiserTilCleanup(): Cleanup  = omsorgsdagerMeldingKonfigurertMapper().readValue(data.rawJson)
@@ -63,7 +67,7 @@ internal fun Any.serialiserTilData() = Data(omsorgsdagerMeldingKonfigurertMapper
 class SerDes : Serializer<TopicEntry>, Deserializer<TopicEntry> {
     override fun configure(configs: MutableMap<String, *>?, isKey: Boolean) {}
     override fun close() {}
-    override fun serialize(topic: String, entry: TopicEntry): ByteArray = when (topic == Topics.K9_RAPID_V2.name) {
+    override fun serialize(topic: String, entry: TopicEntry): ByteArray = when (topic == Topics.K9_RAPID_V2.name ||topic == Topics.K9_DITTNAV_VARSEL.name ) {
         true -> entry.data.rawJson.toByteArray()
         false -> entry.rawJson.toByteArray()
     }
